+adding tdController finished

+++ code/server/controllers/testDescriptorController.js
@@ -18,9 +18,9 @@ const getTestDescriptorById = async (req, res, err) => {
 
 //POST: New Test Descriptor
 const addTestDescriptor =  async (req, res, err) => {
-    /*
-        @To-do:
-     */
+    if (isNaN(req.params.idSKU)) return res.status(422).send("validation of request body failed");
+    // let skuID = await db.getSKU(req.body.idSKU).catch(err);
+    // if (!skuID) return res.status(404).send("no sku associated idSKU");
     await db.newTestDescriptorTable().catch(err);
     let nTD = new TestDescriptor(
         req.body.name,
@@ -34,20 +34,26 @@ const addTestDescriptor =  async (req, res, err) => {
 
 //PUT: Update Test Descriptor
 const updateTestDescriptor = async (req, res, err) => {
+    if (isNaN(req.body.newIdSKU) || isNaN(req.params.id))
+        return res.status(422).send("validation of request body or of id failed");
+    // let skuID = await db.getSKU(req.body.idSKU).catch(err);
+    // if (!skuID) return res.status(404).send("no sku associated idSKU");
     let nTD = new TestDescriptor(
         req.body.newName,
         req.body.newProcedureDescription,
         req.body.newIdSKU);
     let ret = await db.updateTestDescriptor(nTD, req.params.id).catch(err);
-    return (ret ? res.status(201).json({ "message": `${ ret }` }) :
-        res.status(503).json({ "message": "Generic Error." }));
+    return (ret ? res.status(201).json(req.body) :
+        res.status(503).send("Generic Error."));
 }
 
 //DELETE: Delete Test Descriptor
 const deleteTestDescriptor = async(req, res, err) => {
+    if (isNaN(req.params.id))
+        return res.status(422).send("validation of request body or of id failed");
     let ret = await db.deleteTestDescriptor(req.params.id).catch(err);
-    return ( ret ? res.status(204).json({ "message": "Deleted Successfully." }) :
-        res.status(503).json({ "message": "Generic Error." }));
+    return ( ret ? res.status(204).send("Deleted Successfully.") :
+        res.status(404).send("no test descriptor associated id or no sku associated to IDSku"));
 }
 
 module.exports = {
